+++ R/arrange.R
@@ -8,32 +8,30 @@
 #' @export
 #'
 #' @examples
-#' example_dt <- data.table::data.table(
+#' example_dt <- tidytable(
 #'   a = c(1,2,3),
 #'   b = c(4,5,6),
 #'   c = c("a","a","b"))
 #'
 #' example_dt %>%
-#'   arrange.(a, -c)
+#'   arrange.(c, -a)
+#'
+#' example_dt %>%
+#'   arrange.(c, desc.(a))
 arrange. <- function(.data, ...) {
   UseMethod("arrange.")
 }
 
 #' @export
-arrange..tidytable <- function(.data, ...) {
+arrange..data.frame <- function(.data, ...) {
+
+  .data <- as_tidytable(.data)
 
   dots <- enexprs(...)
 
   eval_expr(
     .data[order(!!!dots)]
   )
-}
-
-#' @export
-arrange..data.frame <- function(.data, ...) {
-  .data <- as_tidytable(.data)
-
-  arrange.(.data, ...)
 }
 
 #' @export