+++ R/count.R
@@ -1,18 +1,16 @@
 #' Count observations by group
 #'
 #' @description
-#' Returns row counts of the dataset. If bare column names are provided, `dt_count()` returns counts by group.
-#'
-#' Supports enhanced selection
+#' Returns row counts of the dataset. If bare column names are provided, `count.()` returns counts by group.
 #'
 #' @param .data A data.frame or data.table
-#' @param ... Columns to group by
+#' @param ... Columns to group by. `tidyselect` compatible.
 #'
 #' @export
 #' @md
 #'
 #' @examples
-#' example_df <- data.table::data.table(
+#' example_df <- tidytable(
 #'   x = 1:3,
 #'   y = 4:6,
 #'   z = c("a", "a", "b"))
@@ -30,20 +28,15 @@ count. <- function(.data, ...) {
 }
 
 #' @export
-count..tidytable <- function(.data, ...) {
+count..data.frame <- function(.data, ...) {
+
+  .data <- as_tidytable(.data)
 
   by <- dots_selector_by(.data, ...)
 
   eval_expr(
     .data[, list(N = .N), by = !!by]
   )
-}
-
-#' @export
-count..data.frame <- function(.data, ...) {
-  .data <- as_tidytable(.data)
-
-  count.(.data, ...)
 }
 
 #' @export