+++ R/get_dummies.R
@@ -3,22 +3,23 @@
 #' @description
 #' Convert character and factor columns to dummy variables
 #'
-#' Supports enhanced selection
-#'
 #' @param .data A data.frame or data.table
-#' @param cols If NULL makes dummy variables for all character & factor columns. Or user can input a single column or a vector of unquoted columns to dummify
+#' @param cols A single column or a vector of unquoted columns to dummify.
+#' Defaults to all character & factor columns using `c(is.character, is.factor)`.
+#' `tidyselect` compatible.
 #' @param prefix TRUE/FALSE - If TRUE, a prefix will be added to new column names
 #' @param prefix_sep Separator for new column names
 #' @param drop_first TRUE/FALSE - If TRUE, the first dummy column will be dropped
 #' @param dummify_na TRUE/FALSE - If TRUE, NAs will also get dummy columns
 #'
+#' @md
 #' @export
 #'
 #' @examples
-#' test_df <- data.table::data.table(
-#'   col1 = c(letters[1:3], NA),
-#'   col2 = as.factor(c(letters[3:1], NA)),
-#'   var1 = rnorm(7,0,1))
+#' test_df <- tidytable(
+#'   col1 = c("a", "b", "c", NA),
+#'   col2 = as.factor(c("a", "b", NA, "d")),
+#'   var1 = rnorm(4,0,1))
 #'
 #' # Automatically does all character/factor columns
 #' test_df %>%
@@ -32,58 +33,38 @@
 #' test_df %>%
 #'   get_dummies.(c(col1, col2))
 #'
+#' # Can drop certain columns using
+#' test_df %>%
+#'   get_dummies.(c(is.character, -col2))
+#'
 #' test_df %>%
 #'   get_dummies.(prefix_sep = ".", drop_first = TRUE)
 #'
 #' test_df %>%
 #'   get_dummies.(c(col1, col2), dummify_na = FALSE)
 get_dummies. <- function(.data,
-                           cols = NULL,
-                           prefix = TRUE,
-                           prefix_sep = "_",
-                           drop_first = FALSE,
-                           dummify_na = TRUE) {
+                         cols = c(is.character, is.factor),
+                         prefix = TRUE,
+                         prefix_sep = "_",
+                         drop_first = FALSE,
+                         dummify_na = TRUE) {
   UseMethod("get_dummies.")
 }
 
 #' @export
 get_dummies..data.frame <- function(.data,
-                                      cols = NULL,
-                                      prefix = TRUE,
-                                      prefix_sep = "_",
-                                      drop_first = FALSE,
-                                      dummify_na = TRUE) {
+                                    cols = c(is.character, is.factor),
+                                    prefix = TRUE,
+                                    prefix_sep = "_",
+                                    drop_first = FALSE,
+                                    dummify_na = TRUE) {
 
   .data <- as_tidytable(.data)
-  cols <- enexpr(cols)
-
-  get_dummies.(.data, cols = !!cols,
-                 prefix = prefix, prefix_sep = prefix_sep,
-                 drop_first = drop_first, dummify_na = dummify_na)
-}
-
-#' @export
-get_dummies..tidytable <- function(.data,
-                                      cols = NULL,
-                                      prefix = TRUE,
-                                      prefix_sep = "_",
-                                      drop_first = FALSE,
-                                      dummify_na = TRUE) {
-
   .data <- shallow(.data)
+
   cols <- enexpr(cols)
 
-  if (is.null(cols)) {
-    # If NULL, select all character & factor cols
-    data_names <- names(.data)
-
-    chr_cols <- data_names[map_lgl.(.data, is.character)]
-    fct_cols <- data_names[map_lgl.(.data, is.factor)]
-
-    cols <- syms(c(chr_cols, fct_cols))
-  } else {
-    cols <- vec_selector(.data, !!cols)
-  }
+  cols <- vec_selector(.data, !!cols)
 
   for (col in cols) {