+++ R/pivot_longer.R
@@ -6,15 +6,15 @@
 #' decreasing the number of rows. The inverse transformation is
 #' \code{pivot_longer.()}. Syntax based on the \code{tidyr} equivalents.
 #'
-#' Supports enhanced selection
-#'
 #' @param .data The data table to pivot longer
-#' @param cols Vector of bare column names. Can add/drop columns.
+#' @param cols Vector of bare column names. Can add/drop columns. `tidyselect` compatible.
 #' @param names_to Name of the new "names" column. Must be a string.
 #' @param values_to Name of the new "values" column. Must be a string.
 #' @param values_drop_na If TRUE, rows will be dropped that contain NAs.
 #' @param ... Additional arguments to pass to `melt.data.table()`
 #'
+#' @export
+#' @md
 #'
 #' @examples
 #' example_dt <- data.table::data.table(
@@ -27,8 +27,6 @@
 #'
 #' example_dt %>%
 #'   pivot_longer.(cols = -z, names_to = "stuff", values_to = "things")
-#'
-#' @export
 pivot_longer. <- function(.data,
                           cols = everything.(),
                           names_to = "name",
@@ -39,12 +37,14 @@ pivot_longer. <- function(.data,
 }
 
 #' @export
-pivot_longer..tidytable <- function(.data,
-                                    cols = everything.(),
-                                    names_to = "name",
-                                    values_to = "value",
-                                    values_drop_na = FALSE,
-                                    ...) {
+pivot_longer..data.frame <- function(.data,
+                                     cols = everything.(),
+                                     names_to = "name",
+                                     values_to = "value",
+                                     values_drop_na = FALSE,
+                                     ...) {
+
+  .data <- as_tidytable(.data)
 
   names <- colnames(.data)
   cols <- enexpr(cols)
@@ -67,23 +67,6 @@ pivot_longer..tidytable <- function(.data,
        value.factor = FALSE))
 }
 
-#' @export
-pivot_longer..data.frame <- function(.data,
-                                       cols = everything.(),
-                                       names_to = "name",
-                                       values_to = "value",
-                                       values_drop_na = FALSE,
-                                       ...) {
-  .data <- as_tidytable(.data)
-  cols <- enexpr(cols)
-
-  .data <- pivot_longer.(
-    .data, !!cols,
-    names_to = names_to, values_to = values_to,
-    values_drop_na = values_drop_na, ...)
-
-  .data
-}
 
 #' @export
 #' @rdname pivot_longer.