+++ R/pivot_wider.R
@@ -5,14 +5,19 @@
 #' decreasing the number of rows. The inverse transformation is
 #' \code{pivot_longer.()}. Syntax based on the \code{tidyr} equivalents.
 #'
-#' Supports enhanced selection
-#'
 #' @param .data the data table to widen
-#' @param id_cols A set of columns that uniquely identifies each observation. Defaults to all columns in the data table except for the columns specified in \code{names_from} and \code{values_from}. Typically used when you have additional variables that is directly related.
-#' @param names_from A pair of arguments describing which column (or columns) to get the name of the output column (\code{name_from}), and which column (or columns) to get the cell values from (\code{values_from}).
-#' @param values_from A pair of arguments describing which column (or columns) to get the name of the output column (\code{name_from}), and which column (or columns) to get the cell values from (\code{values_from}).
+#' @param id_cols A set of columns that uniquely identifies each observation.
+#' Defaults to all columns in the data table except for the columns specified in \code{names_from} and \code{values_from}.
+#' Typically used when you have additional variables that is directly related.
+#' `tidyselect` compatible.
+#' @param names_from A pair of arguments describing which column (or columns) to get the name of the output column (\code{name_from}),
+#' and which column (or columns) to get the cell values from (\code{values_from}).
+#' `tidyselect` compatible.
+#' @param values_from A pair of arguments describing which column (or columns) to get the name of the output column (\code{name_from}),
+#' and which column (or columns) to get the cell values from (\code{values_from}).
+#' `tidyselect` compatible.
 #' @param names_sep the separator between the names of the columns
-#' @param values_fn hould the data be aggregated before casting? If the formula doesn't identify a single observation for each cell, then aggregation defaults to length with a message.
+#' @param values_fn Should the data be aggregated before casting? If the formula doesn't identify a single observation for each cell, then aggregation defaults to length with a message.
 #'
 #' @examples
 #' example_dt <- data.table::data.table(
@@ -27,6 +32,7 @@
 #'   pivot_wider.(names_from = stuff, values_from = things, id_cols = z)
 #'
 #' @export
+#' @md
 pivot_wider. <- function(.data,
                          names_from = name,
                          values_from = value,
@@ -37,12 +43,14 @@ pivot_wider. <- function(.data,
 }
 
 #' @export
-pivot_wider..tidytable <- function(.data,
-                                   names_from = name,
-                                   values_from = value,
-                                   id_cols = NULL,
-                                   names_sep = "_",
-                                   values_fn = NULL) {
+pivot_wider..data.frame <- function(.data,
+                                    names_from = name,
+                                    values_from = value,
+                                    id_cols = NULL,
+                                    names_sep = "_",
+                                    values_fn = NULL) {
+
+  .data <- as_tidytable(.data)
 
   id_cols <- enexpr(id_cols)
   names_from <- enexpr(names_from)
@@ -83,27 +91,6 @@ pivot_wider..tidytable <- function(.data,
   if (length(id_cols) == 0) .data <- .data[, . := NULL][]
 
   .data
-}
-
-#' @export
-pivot_wider..data.frame <- function(.data,
-                                      names_from = name,
-                                      values_from = value,
-                                      id_cols = NULL,
-                                      names_sep = "_",
-                                      values_fn = NULL) {
-  .data <- as_tidytable(.data)
-  id_cols <- enexpr(id_cols)
-  names_from <- enexpr(names_from)
-  values_from <- enexpr(values_from)
-  values_fn <- enexpr(values_fn)
-
-  pivot_wider.(.data,
-                 names_from = !!names_from,
-                 values_from = !!values_from,
-                 id_cols = !!id_cols,
-                 names_sep = names_sep,
-                 values_fn = !!values_fn)
 }
 
 #' @export