+++ R/relocate.R
@@ -3,10 +3,8 @@
 #' @description
 #' Move a column or columns to a new position
 #'
-#' Supports enhanced selection
-#'
 #' @param .data A data.frame or data.table
-#' @param ... A selection of columns to move
+#' @param ... A selection of columns to move. `tidyselect` compatible.
 #' @param .before Column to move selection before
 #' @param .after Column to move selection after
 #'
@@ -32,9 +30,13 @@ relocate. <- function(.data, ..., .before = NULL, .after = NULL) {
 }
 
 #' @export
-relocate..tidytable <- function(.data, ..., .before = NULL, .after = NULL) {
+relocate..data.frame <- function(.data, ..., .before = NULL, .after = NULL) {
+
+  .data <- as_tidytable(.data)
+
   .before <- enexpr(.before)
   .after <- enexpr(.after)
+  .data <- shallow(.data)
 
   if  (!is.null(.before) && !is.null(.after))
     stop("Must supply only one of `.before` and `.after`")
@@ -42,7 +44,8 @@ relocate..tidytable <- function(.data, ..., .before = NULL, .after = NULL) {
   if (is.null(.before) && is.null(.after))
     .before <- 1
 
-  all_cols_i <- seq_along(names(.data))
+  data_names <- names(.data)
+  all_cols_i <- seq_along(data_names)
   selected_cols_i <- dots_selector_i(.data, ...)
 
   if (!is.null(.before)) {
@@ -60,16 +63,9 @@ relocate..tidytable <- function(.data, ..., .before = NULL, .after = NULL) {
   start_cols_i <- start_cols_i[start_cols_i %notin% selected_cols_i]
   final_order_i <- unique(c(start_cols_i, selected_cols_i, all_cols_i))
 
-  .data[, ..final_order_i]
-}
+  final_order <- data_names[final_order_i]
 
-#' @export
-relocate..data.frame <- function(.data, ..., .before = NULL, .after = NULL) {
-  .data <- as_tidytable(.data)
-  .before <- enexpr(.before)
-  .after <- enexpr(.after)
-
-  relocate.(.data, ..., .before = !!.before, .after = !!.after)
+  setcolorder(.data, final_order)[]
 }
 
 #' @export