+++ R/transmute.R
@@ -1,7 +1,7 @@
 #' Add new variables and drop all others
 #'
 #' @description
-#' Unlike `dt_mutate()`, `transmute.()` keeps only the variables that you create
+#' Unlike `mutate.()`, `transmute.()` keeps only the variables that you create
 #'
 #' @param .data A data.frame or data.table
 #' @param ... Columns to create/modify
@@ -9,7 +9,6 @@
 #'
 #' @md
 #' @examples
-#' # transmute keeps only the variables you create
 #' mtcars %>%
 #'   transmute.(displ_l = disp / 61.0237)
 #'
@@ -19,24 +18,17 @@ transmute. <- function(.data, ..., by = NULL) {
 }
 
 #' @export
-transmute..tidytable <- function(.data, ..., by = NULL) {
-
-  dots <- enexprs(...)
-  by <- enexpr(by)
-  by <- vec_selector_by(.data, !!by)
-
-  eval_expr(
-    .data[, list(!!!dots), by = !!by]
-  )
-}
-
-#' @export
 transmute..data.frame <- function(.data, ..., by = NULL) {
 
   .data <- as_tidytable(.data)
-  by <- enexpr(by)
 
-  transmute.(.data, ..., by = !!by)
+  dots <- enexprs(...)
+  by <- enexpr(by)
+  keep_names <- names(dots)
+
+  .data <- mutate.(.data, ..., by = !!by)
+
+  .data[, ..keep_names]
 }
 
 #' @export