+++ NEWS.md
@@ -1,3 +1,33 @@
+# tidytable 0.4.1
+
+* `tidytable` now utilizes `tidyselect` & `vctrs`
+* Breaking changes:
+  + `get_dummies.()`: `cols` arg defaults to `c(is.character, is.factor)`.
+  The result is the same, but is more consistent with other enhanced selection functions
+  + `mutate_across.()`: `.funs` argument renamed to `.fns` to match dplyr 1.0
+* New functions:
+  + `desc.()`
+  + `unite.()`
+* Functionality improvements:
+  + Paged printing now works in Rmarkdown
+  + `bind_rows.()`: Does name checking and fills missing
+  + `distinct.()`: Added `.keep_all` argument
+  + `fill.()`: Preserves column order when using "by"
+  + `group_split.()`: Added `keep` arg
+  + `nest.()`
+    - Can now unnest multiple columns in one call
+    - data.tables in a list column can now have different ncols
+    and different column order
+  + `select.()`: Now allows column renaming
+  + `transmute.()`: Summary functions (`mean`/`max`/etc.) can now be used
+* Functions with notable speed improvements:
+  + `distinct.()`
+  + `drop_na.()`
+  + `relocate.()`
+* Bug fixes:
+  + `group_split.()`: Now returns a list of tidytables instead of data.tables
+  + `left_join.()`: Now works with "by" columns of different names
+
 # tidytable 0.4.0
 
 * Function rewrite: